fix 'Boost not found' on macOS Sierra + Boost 1.65.1
From the `FindBoost.cmake` file ([findboost]):

Boost_USE_MULTITHREADED  - Set to OFF to use the non-multithreaded
                           libraries ('mt' tag).  Default is ON.

Because Boost_USE_MULTITHREADED was set to OFF, the requirement on
'thread' was looking for libboost_thread.a.

But the Boost 1.65.1 installed by Homebrew only installs
`libboost_thread-mt.a`, not `libboost_thread.a`.

So I had to turn Boost_USE_MULTITHREADED to ON (it is the same
as commenting it out)

[findboost]: https://github.com/Kitware/CMake/blob/master/Modules/FindBoost.cmake